+++ src/api/server.js
@@ -17,13 +17,16 @@ const { routes } = require('./routes');
  */
 module.exports = (logger, api) => {
   let routesPrefix = api.logPrefix || '/';
+  let logFilesPrefix = routesPrefix + '/log-files';
+  const express = require('express');
   // pass express app
   if (api.appExpress && !_.isEmpty(api.appExpress)) {
+    api.appExpress.use(logFilesPrefix, express.static(logger.logDir));
     routes(logger, api.appExpress, routesPrefix);
-  // create express app
-} else if (api.port) {
-    const express = require('express');
+    // create express app
+  } else if (api.port) {
     let app = express();
+    app.use(logFilesPrefix, express.static(logger.logDir));
     routes(logger, app, routesPrefix);
     app.listen(api.port, () => {
       console.log("Logging server is working on ", api.port);